[openwrt][MAC80211][eth][Fix IPSec traffic not entering SW fast path issue]
[Description]
Fix IPSec traffic not entering SW fast path issue.

Without this patch, the IPSec traffic conntrack would not be set to
OFFLOAD due to dscp learning function could not count the encry
direction's packet count to qos counter, then the TCP Bi-Di > 30
packets condition could not be fulfilled, leads to the OFFLOAD
rule could not be set.

[Release-log]
N/A


Change-Id: I75c0b533a53260fd22da129e8baf1b03731ac337