[kernel-6.6][crypto][common][Refactor inside-secure interrupt handler to avoid rcu stall with IPSec]
[Description]
Refactor inside-secure interrupt handler to avoid rcu stall with IPSec.
For crypto complete callbacks triggered by interrupt handler (bottom
half). We need to prevent context switches to avoid deadlocks.
In the past, we use local_bh_disable/enable in the for-loop and only
lock the section of crypto complete callbacks. And it works fine in
kernel 5.4.
However, in kernel 6.6, this may lead to rcu stall. Since
local_bh_enable will invoke do_softirq then context switch in for-loop.
This will make the interrupt handler take lots of time for process and
trigger rcu stall at the end.

[Release-log]
N/A


Change-Id: I429cf14ce48c05b7d45930aa8a24306c65e42982