[openwrt][mt7988][crypto][Add skb headroom size check for IPSec inline mode]
[Description]
Add skb headroom size check for IPSec inline mode.
In xfrm framework, We'll carry  "cdrt_idx" in the skb headroom. So, in
Tx, we can recongnize that this packet is for IPSec inline mode and need
to be sent to EIP197.
Since the skb from CPU may not have enough headroom size for our data
structure. When we set "cdrt_idx", it might modify the values in the
packet header, causing abnormal packet.
With this patch, we'll check whether skb has enough headroom size before
setting "cdrt_idx". If the headroom size is not enough, we'll copy a new
skb with enough headroom size and replace the original one.

[Release-log]
N/A

Change-Id: I09795f2d8f96686bac6ba23f3e56b8ac9ec992b6