[openwrt-24][mt7988][secure-boot][add estimated size for hash/signature in mkimage]
[Description]
Add estimated size for hash/signature in mkimage. mkimage in
u-boot-v2024.07 signs the same image repeatly if the reserved space for
itb is not enough. Sign keys are accessed multiple times for the same
image. This is not suitable when keys are not in the build environment
since keys are under stricter access control in the scenario. Backport
the solution from u-boot-v2025.10-rc1 to estimate a rough size to
prevent using keys for the same operation multiple times.

[Release-log]
N/A

Change-Id: I00cbe1b79c8f2f8b79d5cf1877398a4b6e2a415d