[kernel][common][eth][Refactor the ETH debugfs infrastructure to align with upstream coding style]
[Description]
Refactor the ETH debugfs infrastructure to align with upstream coding
style.

In this change, we are moving all the debug commands to the debugfs
and removing all global variables to make ETH debugfs more aligned
with upstream style.
Additionally, to maintain backward compatibility with legacy usage,
we are also duplicating some legacy debug commands in the procfs.

[Release-log]
N/A


Change-Id: Iead80db2ae5eb7a89ddc1909e8e28cc66ec63ade