[kernel-6.12][mt7988][eth][Fix the issue of firmware loss after toggling the interface up and down for the Aquantia 10G PHY]
[Description]
Fix the issue of firmware loss after toggling the interface up and down
for the Aquantia 10G PHY.

Because a GPIO reset clears the PHY firmware, we should relocate the
firmware load operation to the config_init function. This relocation
will assist in reloading the PHY firmware by the driver.

Without this patch, the PHY firmware will be lost and cannot be reloaded
after a GPIO reset.

[Release-log]
N/A


Change-Id: I5f95ef7c41f6e705a9a888d31b8819a698d74186