[openwrt][mt7988][crypto][Fix out-of-memory issue with EIP197 DDK lookaside mode]
[Description]
Fix out-of-memory issue with EIP197 DDK lookaside mode.
The interrupt of EIP197 can only be triggered by the number of processed
packets. In the past, we set the irq will be triggered with 10 packets.
When processing UDP data packets, kernel softirq will keep receiving
packets and interrupt is kept invoked which lead to out-of-memory.
Thus, we change the number to trigger irq from 10 to 100.

[Release-log]
N/A

Change-Id: Ia154aa93e70cb905942bab32bf84381fee778081